+++ Menu/MainMenu.java
@@ -15,7 +15,6 @@ import java.awt.event.KeyListener;
 import java.awt.event.KeyEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
-import java.util.ArrayList;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JTextField;
@@ -32,12 +31,6 @@ public class MainMenu extends JFrame {
 	private Menu mainMenuPanel;
 	private Menu optionPanel;
 	private BackgroundPanel backgroundPanel;
-
-	//attributes relative to the option menu
-	private KeyBindingMenu redPlayerBindings;
-	private KeyBindingMenu bluePlayerBindings;
-	private Menu saveQuitOptionsPanel;
-	private Menu saveFailedPanel;
 	
 	/**Contient l'affichage du jeu */
 	private BoardGraphism boardGraphism;
@@ -54,17 +47,13 @@ public class MainMenu extends JFrame {
 	private Menu endGamePanel;
 
 	//boolean handling escape panel
-	private Boolean isDisplayingMainMenu;
 	private Boolean isEscapePanelDisplayed = false;
-
-	//boolean to handle unsaved changes in options
-	private Boolean unsavedChanges = false;
 
 	//the frame displaying all the stuff
 	private JFrame frame;
 
-	//all the KeySelectingPanels
-	private ArrayList<KeySelectingPanel> allKeySelectingPanels = new ArrayList<KeySelectingPanel>();
+	//permet la navigation
+	private Menu selectedMenu;
 
 	public MainMenu(JFrame frame) {		
 		this.frame = frame;
@@ -87,8 +76,6 @@ public class MainMenu extends JFrame {
 
 	/** starts the local board and sets the frame to display it */
 	public void startLocalGame() {
-		isDisplayingMainMenu = false;
-
 		//init game
 		board.initGame();
 		boardIO = new BoardLocal(board);
@@ -120,8 +107,6 @@ public class MainMenu extends JFrame {
 
 	/** starts the online board and sets the frame to display it */
 	public void joinOnlineGame(String address) {
-		isDisplayingMainMenu = false;
-
 		//init game, shouldn't be necessary (if not included the platforms are not displayed)
 		board.initGame();
 
@@ -152,31 +137,27 @@ public class MainMenu extends JFrame {
 		backgroundPanel.removeAll();
 		backgroundPanel.add(backgroundPanel.getLabel());
 		backgroundPanel.add(mainMenuPanel);
-		isDisplayingMainMenu = true;
+		selectedMenu = mainMenuPanel;
 		reloadMenuDisplay();
 	}
 
 	public void createMenus() {
 		mainMenuPanel = new Menu(backgroundPanel, mainMenuPanel, frame);
 
-		optionPanel = new Menu(backgroundPanel, mainMenuPanel, frame);
+		optionPanel = new KeyOptionMenu(backgroundPanel, mainMenuPanel, frame);
 		multiplayerPanel = new Menu(backgroundPanel, mainMenuPanel, frame);
 
-		saveFailedPanel = new Menu(backgroundPanel, optionPanel, frame);
 		joinMultiplayerGamePanel = new Menu(backgroundPanel, multiplayerPanel, frame);
 		connectionErrorPanel = new Menu(backgroundPanel, joinMultiplayerGamePanel, frame);
 		playerLeftPanel = new Menu(backgroundPanel, mainMenuPanel, frame);
 
 		escapePanel = new Menu();
-		saveQuitOptionsPanel = new Menu();
 		createMultiplayerGamePanel = new Menu();
 		endGamePanel = new Menu();
 
 		createMainMenuPanel();
 
 		createKeyBindingMenu();
-		createSaveQuitOptionsPanel();
-		createSaveFailedPanel();
 
 		createMultiplayerPanel();
 		createCreateMultiplayerGamePanel();
@@ -212,89 +193,11 @@ public class MainMenu extends JFrame {
 		mainMenuPanel.setOrder(true);
 	}
 
-	/** sets the binding in the bindingMenus to default or personalized bindings according to the existence of personalized bindings */
-	public void setBindings() {
-		redPlayerBindings.setAllBindings(FileFunctions.getPathFileToUse("red"));
-		bluePlayerBindings.setAllBindings(FileFunctions.getPathFileToUse("blue"));
-	}
-
 	/** initiates the components of the menu */
 	public void createKeyBindingMenu() {
 		optionPanel.displayBorder("OPTIONS");
 
-		// panel containing the bindings
-		Menu keyBindingPanel = new Menu();
-		//create each player bindings panel
-		redPlayerBindings = new KeyBindingMenu("Red player bindings", this);
-		bluePlayerBindings = new KeyBindingMenu("Blue player bindings", this);
-		addAllKeySelectingPanels();
-
-		redPlayerBindings.addKeySelectingPanels(FileFunctions.getPathFileToUse("red"), "KeyBindings/redKeyBindingsDefault.txt");
-		bluePlayerBindings.addKeySelectingPanels(FileFunctions.getPathFileToUse("blue"), "KeyBindings/blueKeyBindingsDefault.txt");
-
-		addAllKeySelectingPanels();
-
-		redPlayerBindings.setDimensions();
-		redPlayerBindings.setOrder(true);
-		bluePlayerBindings.setDimensions();
-		bluePlayerBindings.setOrder(true);
-		
-		keyBindingPanel.add(redPlayerBindings);
-		keyBindingPanel.add(bluePlayerBindings);
-		keyBindingPanel.setDimensions();
-		keyBindingPanel.setOrder(false);
-
-		//create a panel to contain the buttons
-		Menu buttonPanel = new Menu();
-
-		/** save bindings */
-		buttonPanel.addNewButton("SAVE", new ActionListener() {
-			public void actionPerformed(ActionEvent arg0) {   
-				saveOptions(false);
-				unsavedChanges = false;
-			}
-		});
-
-		/** back to main menu */
-		buttonPanel.addNewButton("BACK", new ActionListener() {
-			public void actionPerformed(ActionEvent arg0) { 
-				if (!unsavedChanges) {
-					backToMainMenuFromOption();
-				} else {
-					backgroundPanel.remove(optionPanel);
-					backgroundPanel.add(saveQuitOptionsPanel);
-					reloadMenuDisplay();
-				}
-			}
-		});
-
-		/** reset to default bindings */
-		buttonPanel.addNewButton("RESET BINDINGS", new ActionListener() {
-			public void actionPerformed(ActionEvent arg0) {        
-				//check if non default key settings exist and delete those files
-				FileFunctions.deleteNonDefaultBindings();
-				//recreate the menu with default settings 
-				setBindings();
-				unsavedChanges = false;
-			}
-		});
-
-		//make it horizontal
-		buttonPanel.setDimensions();
-		buttonPanel.setOrder(false);
-
-		//add the player's bindings panels to main panel
-		optionPanel.add(keyBindingPanel);
-		//and the buttons at the bottm
-		optionPanel.add(buttonPanel);
-
 		optionPanel.setOrder(true);
-	}
-
-	//create an array list of all KeySelectingPanels
-	public void addAllKeySelectingPanels() {
-		allKeySelectingPanels.addAll(redPlayerBindings.getKeySelectingPanels());
-		allKeySelectingPanels.addAll(bluePlayerBindings.getKeySelectingPanels());
 	}
 
 	/** creates the panel that opens when pressing escape */
@@ -336,49 +239,6 @@ public class MainMenu extends JFrame {
 		}
 		isEscapePanelDisplayed = !isEscapePanelDisplayed;
 		frame.setVisible(true);
-	}
-
-	public void createSaveQuitOptionsPanel() {
-		saveQuitOptionsPanel.displayBorder("WARNING");
-
-		saveQuitOptionsPanel.add(new JLabel("Some changes have not been saved"));
-
-		Menu buttonPanel = new Menu();
-
-		/** resume game */
-		buttonPanel.addNewButton("SAVE AND QUIT", new ActionListener() {
-			public void actionPerformed(ActionEvent arg0) { 
-				backgroundPanel.remove(saveQuitOptionsPanel);
-				backgroundPanel.add(optionPanel);
-				reloadMenuDisplay();
-				saveOptions(true);
-			}
-		});
-
-		/** go back to the main menu */
-		buttonPanel.addNewButton("CANCEL AND QUIT", mainMenuPanel);
-
-		buttonPanel.setDimensions();
-		buttonPanel.setOrder(false);
-
-		// add all the components
-		saveQuitOptionsPanel.add(buttonPanel);
-
-		saveQuitOptionsPanel.setDimensions();
-		saveQuitOptionsPanel.setOrder(true);
-	}
-
-	public void createSaveFailedPanel() {
-		saveFailedPanel.displayBorder("WARNING");
-
-		JLabel info = new JLabel("There's a problem with your bindings");
-		saveFailedPanel.add(info);
-
-		/** resume game */
-		saveFailedPanel.addNewButton("OK");
-
-		saveFailedPanel.setDimensions();		
-		saveFailedPanel.setOrder(true);
 	}
 
 	/** creates the multiplayerJPanel with its component*/
@@ -441,7 +301,7 @@ public class MainMenu extends JFrame {
 		buttonPanel.addNewButton("BACK", new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
 				//displays the mainMenu panel 
-				buttonPanel.backInteraction();
+				buttonPanel.menuInteraction();
 				timer.stop();
 			}
 		});
@@ -493,7 +353,7 @@ public class MainMenu extends JFrame {
 				//displays the multiplayer panel
 				boardIO.handleKeyListeners(frame, false);
 				backgroundPanel.removeAll();
-				playerLeftPanel.backInteraction();
+				playerLeftPanel.menuInteraction();
 			}
 		});
 
@@ -521,7 +381,7 @@ public class MainMenu extends JFrame {
 		//back to main menu
 		buttonPanel.addNewButton("BACK TO MAIN MENU", new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
-				buttonPanel.backInteraction();
+				buttonPanel.menuInteraction();
 				boardIO.exitGame();
 			}
 		});
@@ -561,50 +421,6 @@ public class MainMenu extends JFrame {
 		frame.setVisible(true);
 	}
 
-
-	/** saves options if they are valid */
-	public void saveOptions(Boolean back) {
-		//saves the current (the ones being displayed) keyBindings if they are all unique
-		if (checkUnicity()) {
-			KeyBindings redBindings = redPlayerBindings.getCurrentKeyBindings();
-			KeyBindings blueBindings = bluePlayerBindings.getCurrentKeyBindings();
-			FileFunctions.saveObject(redBindings, "KeyBindings/redKeyBindings.txt");
-			FileFunctions.saveObject(blueBindings, "KeyBindings/blueKeyBindings.txt");
-			if (back) {
-				backToMainMenuFromOption();
-			}
-		} else {
-			backgroundPanel.remove(optionPanel);
-			backgroundPanel.add(saveFailedPanel);
-			reloadMenuDisplay();
-		}
-	}
-
-
-	/** check if every binding is unique */
-	public boolean checkUnicity() {
-		for (KeySelectingPanel keySelectingPanel : allKeySelectingPanels) {
-			for (KeySelectingPanel keySelectingPanel2 : allKeySelectingPanels) {
-				if (keySelectingPanel != keySelectingPanel2 && keySelectingPanel.getCurrentKeyBinding().getKeyValue() == keySelectingPanel2.getCurrentKeyBinding().getKeyValue()) {
-					return false;
-				}
-			}
-			//there must be a way to decrease complexity by half and avoid testing twice
-		}
-		return true;
-	}
-
-
-	/** displays main menu and cancel unsaved bindings changes */
-	public void backToMainMenuFromOption() {
-		//all options are saved
-		unsavedChanges = false;
-		//displays the main menu again 
-		displayMainMenu();
-		//cancels changes if they are not saved
-		setBindings();
-	}
-
 	/** Find public IP address */
 	public String getPublicIPAddress() {
 		try { 
@@ -619,20 +435,8 @@ public class MainMenu extends JFrame {
 		} 
 	}
 
-	public void setUnsavedChanges(Boolean bool) {
-		unsavedChanges = bool;
-	}
-
 	public Board getBoard() {
 		return board;
-	}
-
-	public KeyBindingMenu getRedPlayerBindingMenu() {
-		return redPlayerBindings;
-	}
-
-	public KeyBindingMenu getBluePlayerBindingMenu() {
-		return bluePlayerBindings;
 	}
 
 	public class ClientRelatedKeyListener implements KeyListener {
@@ -648,11 +452,7 @@ public class MainMenu extends JFrame {
 
 			//escape
 			if (code == 27) {
-				if (frame.getContentPane() == backgroundPanel && !isDisplayingMainMenu) {
-					backToMainMenuFromOption();
-				} else if (frame.getContentPane() == boardGraphism) {
-					toggleEscapePanel();
-				}
+				selectedMenu.menuInteraction();
 			}
 		}
